Change nav-bar to show logged-in as user after success
