+++ routes/attempts.js
@@ -1,32 +1,16 @@
 const express = require('express');
 const router  = express.Router();
-const db = require('../db/connection');
-
-const getAttemptsForUserID = function(userid) {
-  return db
-  .query(`
-  SELECT quizzes.title, attempts.grade
-  FROM attempts
-  JOIN users ON attempts.user_id=users.id
-  JOIN quizzes ON attempts.quiz_id=quizzes.id
-  WHERE users.id = $1;
-  `,[userid])
-  .then(res => {
-    console.log(res.rows);
-    return res.rows;
-  })
-  .catch(err => console.error(err.message));
-};
+const database = require("../db/queries/quiz");
 
 router.get('/:userId', (req, res) => {
   // Get the userId from the URL parameter
   const userID = req.params.userId || req.session.userId;
   console.log(userID);
-  getAttemptsForUserID(userID)
+  database.getAttemptsForUserID(userID)
   .then((attempts) => {
-   
+    const user = req.session.user;
     // Pass attempts data to the template
-    res.render('attempts', {attempts});
+    res.render('attempts', {attempts, user});
   })
   .catch((err) => {
     console.error(err.message);
