+++ views/attempts.ejs
@@ -9,24 +9,36 @@
   <link rel="stylesheet" href="/vendor/border-box.css" />
   <link rel="stylesheet" href="/styles/main.css" />
   <link rel="stylesheet" href="/styles/layout.css" />
+  <link rel="stylesheet" href="/styles/attempts.css" />
   <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/skeleton/2.0.4/skeleton.min.css" integrity="sha512-EZLkOqwILORob+p0BXZc+Vm3RgJBOe1Iq/0fiI7r/wJgzOFZMlsqTa29UEl6v6U6gsV4uIpsNZoV32YZqrCRCQ==" crossorigin="anonymous" referrerpolicy="no-referrer" />
   <script src="/vendor/jquery-3.0.0.js"></script>
   <script defer src="/scripts/app.js"></script>
 </head>
 
 <body>
-  <header>
-   
-  </header>
-
-  <div class="container">
-    <ul class="Attempt-list">
+  <div class="attempts-container">
+    <h1>Attempts for <%= attempts[0].name %></h1>
+    <table>
+      <table>
+        <thead>
+          <tr class="center">
+            <th>Title</th>
+            <th>Grade</th>
+          </tr>
+        </thead>
+      <tbody>
       <% for (let i = 0; i < attempts.length; i++) { %>
-        <li>
-          <%= attempts[i].title %> <%= attempts[i].grade %>
-        </li>
+        <tr>
+          <td class="center">
+            <%= attempts[i].title %>
+          </td>
+          <td class="center">
+            <%= attempts[i].grade %>
+          </td>
+        </tr>
       <% } %>
-    </ul>
+    </tbody>
+  </table>
   </div>
 </body>
