+++ views/index.ejs
@@ -27,7 +27,7 @@
         <div>
         <ul id="navbar-menu">
           <li><a href="/">Home</a></li>
-          <li><a href="/createQuiz">Create Quiz</a></li>
+          <li><a href="/quiz-settings">Create Quiz</a></li>
           <li><a href="/quizzes">Public Quizzes</a></li>
           <li id="login-item"><a href="/login" id="login-link">Login</a></li>
         </ul>
